refactor: move files and fix references
